Undo double fix of compiler warnings.
git-svn-id: svn+ssh://svn.gnucash.org/repo/gnucash/trunk@11954 57a11ea4-9604-0410-9ed3-97b8803252fd

+++ src/core-utils/gnc-gkeyfile-utils.c
@@ -123,10 +123,10 @@ gnc_key_file_save_to_file (const gchar *filename,
     if (error) {
       *error = g_error_new(G_FILE_ERROR, g_file_error_from_errno(errno), 
 			   "File %s truncated (provided %d, written %zd)",
-			   filename, length, (int) written);
+			   filename, length, written);
     } else {
       g_critical("File %s truncated (provided %d, written %zd)",
-	      filename, length, (int) written);
+	      filename, length, written);
     }
     /* Ignore any error */
     close(fd);
